+++ example/lib/main.dart
@@ -1,5 +1,4 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_cache_manager/flutter_cache_manager.dart';
 import 'package:flutter_localizations/flutter_localizations.dart';
 import 'package:flutter_toolbox/flutter_toolbox.dart';
 import 'package:flutter_toolbox/generated/l10n.dart' as toolbox;
@@ -7,15 +6,6 @@ import 'package:flutter_toolbox/generated/l10n.dart' as toolbox;
 import 'paginated_list_view_example.dart';
 
 void main() => runApp(MyApp());
-
-class MyCM extends BaseCacheManager {
-  MyCM(String cacheKey) : super(cacheKey);
-
-  @override
-  Future<String> getFilePath() {
-    return null;
-  }
-}
 
 class MyApp extends StatefulWidget {
   @override