+++ test/flutter_toolbox_test.dart
@@ -1,1 +1,8 @@
-void main() {}
+import 'package:flutter_test/flutter_test.dart';
+
+void main() {
+  test('String.split() splits the string on the delimiter', () {
+    var string = 'foo,bar,baz';
+    expect(string.split(','), equals(['foo', 'bar', 'baz']));
+  });
+}